dendrogram: Move tests
